[refactor] CalendarViewModel을 HomeViewModel과 병합
