+++ DailyQuest/DailyQuest/Presentation/Home/ViewModel/HomeViewModel.swift
@@ -12,21 +12,28 @@ import RxCocoa
 
 final class HomeViewModel {
     private let questUseCase: QuestUseCase
+    private var calendarUseCase: CalendarUseCase
     
-    init(questUseCase: QuestUseCase) {
+    init(questUseCase: QuestUseCase, calendarUseCase: CalendarUseCase) {
         self.questUseCase = questUseCase
+        self.calendarUseCase = calendarUseCase
     }
     
     struct Input {
         let viewDidLoad: Observable<Date>
         let itemDidClicked: Observable<Quest>
+        
+        let dragEventInCalendar: Observable<CalendarView.ScrollDirection>
     }
     
     struct Output {
         let data: Driver<[Quest]>
+        
+        let currentMonth: Observable<Date?>
+        let displayDays: Driver<[[DailyQuestCompletion]]>
     }
     
-    func transform(input: Input) -> Output {
+    func transform(input: Input, disposeBag: DisposeBag) -> Output {
         
         let updated = input
             .itemDidClicked
@@ -47,6 +54,37 @@ final class HomeViewModel {
             .flatMap(questUseCase.fetch(by:))
             .asDriver(onErrorJustReturn: [])
         
-        return Output(data: data)
+        input
+            .viewDidLoad
+            .subscribe(onNext: { [weak self] _ in
+                self?.calendarUseCase.setupMonths()
+            })
+            .disposed(by: disposeBag)
+        
+        input
+            .dragEventInCalendar
+            .subscribe(onNext: { [weak self] direction in
+                switch direction {
+                    case .prev:
+                        self?.calendarUseCase.fetchLastMontlyCompletion()
+                    case .none:
+                        break
+                    case .next:
+                        self?.calendarUseCase.fetchNextMontlyCompletion()
+                }
+            })
+            .disposed(by: disposeBag)
+        
+        let currentMonth = calendarUseCase
+            .currentMonth
+            .asObserver()
+        
+        let displayDays = calendarUseCase
+            .completionOfMonths
+            .asDriver(onErrorJustReturn: [[], [], []])
+        
+        return Output(data: data,
+                      currentMonth: currentMonth,
+                      displayDays: displayDays)
     }
 }
